move tests from gear-core to gprimitives
