+++ core/src/ids.rs
@@ -156,68 +156,3 @@ pub mod prelude {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    extern crate alloc;
-
-    use super::{prelude::*, CodeId, ProgramId};
-    use alloc::format;
-
-    #[test]
-    fn formatting_test() {
-        let code_id = CodeId::generate(&[0, 1, 2]);
-        let id = ProgramId::generate_from_user(code_id, &[2, 1, 0]);
-
-        // `Debug`/`Display`.
-        assert_eq!(
-            format!("{id:?}"),
-            "0x6a519a19ffdfd8f45c310b44aecf156b080c713bf841a8cb695b0ea5f765ed3e"
-        );
-        // `Debug`/`Display` with precision 0.
-        assert_eq!(format!("{id:.0?}"), "0x..");
-        // `Debug`/`Display` with precision 1.
-        assert_eq!(format!("{id:.1?}"), "0x6a..3e");
-        // `Debug`/`Display` with precision 2.
-        assert_eq!(format!("{id:.2?}"), "0x6a51..ed3e");
-        // `Debug`/`Display` with precision 4.
-        assert_eq!(format!("{id:.4?}"), "0x6a519a19..f765ed3e");
-        // `Debug`/`Display` with precision 15.
-        assert_eq!(
-            format!("{id:.15?}"),
-            "0x6a519a19ffdfd8f45c310b44aecf15..0c713bf841a8cb695b0ea5f765ed3e"
-        );
-        // `Debug`/`Display` with precision 30 (the same for any case >= 16).
-        assert_eq!(
-            format!("{id:.30?}"),
-            "0x6a519a19ffdfd8f45c310b44aecf156b080c713bf841a8cb695b0ea5f765ed3e"
-        );
-        // `Debug`/`Display` with sign + (vara address).
-        assert_eq!(
-            format!("{id:+}"),
-            "kGhwPiWGsCZkaUNqotftspabNLRTcNoMe5APCSDJM2uJv6PSm"
-        );
-        // `Debug`/`Display` with width (custom address, 42 means substrate).
-        assert_eq!(
-            format!("{id:42}"),
-            "5EU7B2s4m2XrgSbUyt8U92fDpSi2EtW3Z3kKwUW4drZ1KAZD"
-        );
-        // Alternate formatter.
-        assert_eq!(
-            format!("{id:#}"),
-            "ActorId(0x6a519a19ffdfd8f45c310b44aecf156b080c713bf841a8cb695b0ea5f765ed3e)"
-        );
-        // Alternate formatter with precision 2.
-        assert_eq!(format!("{id:#.2}"), "ActorId(0x6a51..ed3e)");
-        // Alternate formatter with sign + (vara address).
-        assert_eq!(
-            format!("{id:+#}"),
-            "ActorId(kGhwPiWGsCZkaUNqotftspabNLRTcNoMe5APCSDJM2uJv6PSm)"
-        );
-        // Alternate formatter with width (custom address, 42 means substrate).
-        assert_eq!(
-            format!("{id:#42}"),
-            "ActorId(5EU7B2s4m2XrgSbUyt8U92fDpSi2EtW3Z3kKwUW4drZ1KAZD)"
-        );
-    }
-}
