+++ core/src/program.rs
@@ -140,15 +140,6 @@ mod tests {
     }
 
     #[test]
-    /// Test that ProgramId's `try_from(bytes)` constructor causes panic
-    /// when the argument has the wrong length
-    fn program_id_from_slice_error_implementation() {
-        let bytes = "foobar";
-        let result: Result<ProgramId, _> = bytes.as_bytes().try_into();
-        assert!(result.is_err());
-    }
-
-    #[test]
     /// Test static pages.
     fn program_memory() {
         let wat = r#"
